+++ Backend/database.sql
@@ -29,5 +29,6 @@ CREATE TABLE "users" (
   "recovery" varchar(9)[],
   "credit_card" varchar(64),
   "crypto_wallet" varchar(64),
+  "api_access" boolean default false,
   "api_key" varchar(32)
 );