+++ Frontend/src/routes/settings/+page.server.ts
@@ -1,22 +1,15 @@
 import type {PageServerLoad, Actions} from './$types';
 import type {Settings} from '$type';
+import {fetchApi} from '$lib';
 
-export const load: PageServerLoad = () => {
-  return {
-    userSettings: {
-      has2FA: true,
-      hasApiAccess: true,
-      recoveryCodes: [458962137, 856239147, 125896347],
-      apiKey: 'hfzjfzzefzh',
-      billing: 'hfezehfezjz',
-    } satisfies Settings,
-  };
+export const load: PageServerLoad = async () => {
+  const response = await fetchApi('/settings/', 'GET');
+  return {settings: response};
 };
 
 export const actions: Actions = {
   username: async ({request}) => {},
   password: async ({request}) => {},
-  toggleOtp: async ({request}) => {},
   otp: async ({request}) => {},
   recovery: async ({request}) => {},
   toggleApi: async ({request}) => {},